Prevent splices over part-heads for `splice_style = "calls"`
