+++ monument/test/cases/splice-over-part-head.toml
@@ -1,3 +1,5 @@
+# This tests that splices over part heads are counted.  For example `CY[H]x3` should have a weight
+# of 5 (splices), not 3 (as you'd get if you discounted splices over the part heads).
 length = { min = 0, max = 800 }
 part_head = "1342"
 course_heads = ["*5678"]
